QmlDesigner: Fix the position of placeholder in FilterComboBox
The AutoCompleteString view had a bad position. The text property of the
TextMetrics was wrong, which caused the placeholder to have bad position
since the position was calculated with wrong text.

In addition, I prevented showing the placeholder when the root comboBox
is not opened, because otherwise it may show text when no item is
selected.

The main task was QDS-7662, which is now divided.

Task-number: QDS-7912
Change-Id: Ib937d923191ec3c544dae3259774cd2bb5b7adb0
Reviewed-by: Henning Gründl <henning.gruendl@qt.io>