Debugger: Fix interrupting with cdb
It turns out sending ctrl+c events to cdb results in inconsistent
behavior when trying to interrupt with the range from interrupting twice
to not at all. Use the previous method of using signal operation for
local debugging, and use the ctrl+c event only when cdb is attached to a
remote server.

Fixes: QTCREATORBUG-28279
Change-Id: Iccd2016685ba707b375aebfd88eccc253dde1d1d
Reviewed-by: Christian Stenger <christian.stenger@qt.io>