McuSupport: Hardcode the title of the groupbox
The main McuPackage does not provide a title anymore. It was removed in
df92e79da4a6d10ac64d011faeabf4ffbfaa9ec2.
Since then, the title was empty and the second group box showed a faulty
text. This change hardcodes the product name as it is not going to change
very often.

Task-number: QTCREATORBUG-28302
Change-Id: I6180579c9dce2f2f8aa939fb99088b4867d699fc
Reviewed-by: Yasser Grimes <yasser.grimes@qt.io>
Reviewed-by: hjk <hjk@qt.io>