Sections could be non-collapsible
A property named collapsible is considered for Sections. The default
value is true for this property.
The section would be affected by the user collapse request if the
collapsible is true (Single collapse by clicking, and colapseAll
request). Otherwise, the section will not be affected by user
requests.

Task-number: QDS-7527
Change-Id: Ic4c9d0489d4b8c9ec90a0c71c755fa4c52935d39
Reviewed-by: Miikka Heikkinen <miikka.heikkinen@qt.io>
Reviewed-by: Mahmoud Badri <mahmoud.badri@qt.io>