ImageViewer: Improve tool bar size and behavior
Merge the various export functionality into one tool button instead of
three, for this not-so-often used functionality.

Put the actions into a QToolBar, so if the editor area is made so small
that not all actions fit anymore, the tool bar gets the standard
extension button for accessing the "overflowing" ones.

Fixes: QTCREATORBUG-28309
Change-Id: I2e1fcee9414038aca49648f18f61e1f15ecf3e5a
Reviewed-by: <github-actions-qt-creator@cristianadam.eu>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Alessandro Portale <alessandro.portale@qt.io>