McuSupport: Indicate the usage of QtCreator9+ for QtMCUs 2.3 and above
Starting from QtMCUs 2.3 legacy code is not used in McuSupport plugin
and the refactored code is used instead.
Trying to use 2.3 in older versions of QtCreator will result is using
incomplete refactored code causing errors when generating kits.

This commit will indicate to the user to choose the correct version of
QtCreator for newer QtMCUs releases.

Task-number: QTCREATORBUG-28286
Change-Id: Ib3934ef50f9cee81f1fac8a36674473fc2169ef5
Reviewed-by: Rainer Keller <Rainer.Keller@qt.io>
Reviewed-by: Sivert Krøvel <sivert.krovel@qt.io>
Reviewed-by: hjk <hjk@qt.io>