ProjectExplorer: Ensure working dir is reachable
Previously the working directory might not be reachable by
the executable.

This change fixes that by calling "ensureReachable", and
changing the path to an on-device path of the exectuable.

This also fixes the path when the Host is Windows, and
the build device is a unix type system by changing the paths
from e.g. "c:/..." to "/c/..."

Change-Id: I6c86e16c867bb654b6c2a4951d9f62fdb879298e
Reviewed-by: Alessandro Portale <alessandro.portale@qt.io>