QmlDesigner: Delay project preview generation by 10 seconds
This way the preview generation does not block opening the project
and the project is fully loaded. If the preview is null
we store this in the data base and wait one hour until we
try to regenerate it. Otherwise we might try to render a preview
whenever we switch files.

Task-number: QDS-7992
Change-Id: I80a078ec7b8d418fa1fccd14bc4603b82c36c5e0
Reviewed-by: Marco Bubke <marco.bubke@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>