Support temporarily dragging progress details out of the way
The progress details can cover UI that the user is interested in, and
with e.g. the Clangd indexing, building and update check we have some
long term progress details visible.

It is already possible to hide the details by pressing the tool button,
but another natural reaction is to just drag the UI out of the way.

Add support to drag the progress details to a different position
temporarily. The original "preferred" location is reset either when we
show a new progress detail and we didn't show anything before, or when
the user hides and shows the progress details via the button.

Fixes: QTCREATORBUG-28078
Change-Id: Ie9cecf8b4a34385426a4de3baaf4da92e6387254
Reviewed-by: David Schulz <david.schulz@qt.io>
Reviewed-by: <github-actions-qt-creator@cristianadam.eu>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>