Core: Allow "Remove Folder" in File System View pane
As a CMake user from time to time I have the need to simply remove a
build directory.

This is now possible from the File System pane. Previously one had to
open a Terminal an do there a "rm -rf" or "rmdir /q /s" operation.

The user will be asked before, and there is no possibility of a misclick
and remove something the user didn't consent to.

Fixes: QTCREATORBUG-27331
Change-Id: I61aa42ce6587e46d635d9743b154f2bc9d163b1e
Reviewed-by: Alessandro Portale <alessandro.portale@qt.io>