ProjectExplorer: Use LRU for sorting sessions in menu
Right now, upon start-up of Qt Creator, entries in the Sessions menu are
sorted in reverse order of the time the corresponding files were last
modified on disk, which often (but not always) corresponds to a "least
recently used" order. This order is unchanged during the lifetime of the
current instance of Qt Creator, and potentially changes again after the
next start.
As this can be confusing to users, we now ensure LRU sorting at all
times.

Task-number: QTCREATORBUG-28202
Change-Id: I758371ac1cb968dd3f2e4278fc5066d1e74397c6
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: <github-actions-qt-creator@cristianadam.eu>
Reviewed-by: Eike Ziller <eike.ziller@qt.io>