Prevent a crash on close
The Destructor of the class Model::Private indirecly called
modelAboutToBeDetached which did lead the FormEditor to try
to access the already deleted ModelPrivate.
This patch is a quick-crash-fix that prevents this by moving
the detachAllViews call from the destructor of ModelPrivate
to the destructor of Model.
Note that a clean fix would probably make the access of
ModelPrivate outside of Model impossible but this is outside
the scope of this fix and left for a later exercise.

Fixes: QDS-7587
Change-Id: I0d03423ae6eaf4d86eddba26ec5edb643ea5a51e
Reviewed-by: Marco Bubke <marco.bubke@qt.io>