Boot2Qt: Fix default deployment method
This fix addresses two problems:

1) If no boot2qt device exists yet, no deploy step is created.
This is hard for the user to debug.
2) As the device supports rsync, an RsyncDeployStep would be created.
This is not optimal when the build device is remote, as the fallback
would not check if the target file is up-to-date first.

Therefore this fix adds another possible deploy step that will always
be created if either no device is available, or the build device is a
remote device.

Additionally this change adds an RSync deploy step if neither a remote
build device is set nor a target device.

Change-Id: I8959408e13d690d484ade9e837e7fa65eb106676
Reviewed-by: Samuli Piippo <samuli.piippo@qt.io>
Reviewed-by: <github-actions-qt-creator@cristianadam.eu>