McuSupport: Enable os-specific executable version detection
Currently version detection fails in either Linux or Windows based on
the content of versionDetection.filePattern in the json kit file.

This change enables os-specific filePattern fields:

"versionDetection": {
    "filePattern": {
        "windows": "bin/arm-none-eabi-g++.exe",
        "linux": "bin/arm-none-eabi-g++"
    }
    "executableArgs": "--version",
    "regex": "\\b(\\d+\\.\\d+\\.\\d+)\\b"
}

Simply appending the binary extension with withExecutableSuffix()
is less portable and does not reflect the validationPath field.

For McuPackageDirectoryEntriesVersionDetector and
McuPackageXmlVersionDetector to work correctly, they rely on
filePattern not having os-specific fields.

Also make getOsSpecificValue and parseVersionDetection static.

Change-Id: Ieacf376126043d732eeb5d5a2f4125963022ae76
Reviewed-by: Sivert Krøvel <sivert.krovel@qt.io>
Reviewed-by: Eike Ziller <eike.ziller@qt.io>
Reviewed-by: <github-actions-qt-creator@cristianadam.eu>