McuSupport: Make legacy kit creation match QtCreator 8
The default path for the STM32 Cube Programmer was recently changed for
legacy packages. This means that Qt for MCUs 2.2 kits created with
QtCreator 9 and kits created with QtCreator 8 would have different
paths. This is not ideal as a basis for automatically creating new
kits when changing Qt for MCUs version.

This change makes QtCreator 9 create legacy kits in the same way
QtCreator 8 did.

Change-Id: I8b1fa4a0d673f746edf058aee292e6d42ca2cdcf
Reviewed-by: Yasser Grimes <yasser.grimes@qt.io>
Reviewed-by: <github-actions-qt-creator@cristianadam.eu>
Reviewed-by: Eike Ziller <eike.ziller@qt.io>