McuSupport: Avoid empty string for cmake toolchain vars
At the first run of Qt Creator after its installation, MSVC toolchain
command is set lazily after it's detected, so an empty string can be
set to cmake toolchain variable when a Qt for MCUs kit is created
automatically. It needs be avoided to keep the cmake variables valid.
Even when the command is not updated, it will work correctly with the
default variale (%{Compiler:Executable:(C,CXX)}).

Task-number: QTCREATORBUG-28457
Change-Id: I5de277831e7f1e696f67724193938d6eef7a12dd
Reviewed-by: Sivert Krøvel <sivert.krovel@qt.io>
Reviewed-by: <github-actions-qt-creator@cristianadam.eu>
Reviewed-by: Alessandro Portale <alessandro.portale@qt.io>