QmlDesigner: Fix shaking items when scrubbing the timeline
The reason for this is that when scrubbing the timeline,
the bounding rectangle can change a lot (See bugreport).
If this is the case the pixmap and the bounding rectangle get temporarily
out of sync, which leads to the shaking.
In this patch we add the bounding rectangle to the pixmap command.
The bounding rectangle coming with the pixmap has higher precedence.
This means if there is a pixmap, then the pixmap is always in sync with
the pixmap. If there is no pixmap we use the "original" bounding rectangle.

Task-number: QDS-7828
Change-Id: I40c0b7ed97863b9dca726547927ae1a37f9c415d
Reviewed-by: Tim Jenssen <tim.jenssen@qt.io>