QmlDesigner: Fix states editor none visual items
When using any none visual item as root (e.g. Window) which don't have
a state property the support for adding a state on the root item needs
to be deactivated otherwise QtDS is crashing. The new states editor was
missing that.
On the other hand the use of StatesGroup with none visual items is
supported.

Change-Id: I194913719a13099ce8ac7cde5f78083ef84dc37f
Reviewed-by: <github-actions-qt-creator@cristianadam.eu>
Reviewed-by: Thomas Hartmann <thomas.hartmann@qt.io>