DocumentModel: Fix possible use of deallocated memory
Setting entry to nullptr inside DocumentModelPrivate::addEntry()
sets it only locally inside this function, so the caller of
addEntry will still hold the originally passed pointer. So, if the
"if (previousEntry)" condition was hold, the caller of addEntry
operates on deleted pointer.

In order to fix it we return the valid pointer instead.

Change-Id: I3e4a5c3532ca942c43ab422a238f0a19ebd41794
Reviewed-by: Eike Ziller <eike.ziller@qt.io>