StudioWelcome: Use QQuickWindow on macOS
This fixes flickering on macOS introduced by Qt 6.4.
Unfortuntly keyboard focus does not work on Windows in this case
if the splashscreen flag is set.
Therefore we only use QQuickWindow on macOS.
Simplfiying code and removing unused code.
The variable doNotShowAgain is always true now.
The check in delayedInitialize can be removed,
since we already show a warning during creation.

Task-number: QDS-8113
Change-Id: I8185a58175f4d2eafe92561fbded535d7ae393a2
Reviewed-by: Henning Gründl <henning.gruendl@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>