QtSupport: Warn if Link With Qt failed
When installing Qt Creator via

https://download.qt.io/official_releases/qtcreator/9.0/9.0.0/
qt-creator-opensource-mac-x86_64-9.0.0.dmg

The bundle is not writable / no folders can be created in it.
therefor "Link With Qt" would silently fail, as the button is not
disabled, even if the "canLinkWithQt()" returns false.

This also adds a message box to let the user known if the linking
failed.

Change-Id: Ifbc5d170cf81e898eec3ab8b9f53e1a939652b42
Reviewed-by: Eike Ziller <eike.ziller@qt.io>